BUGFIX: compassomator: additionally use the command program name to match a running process. this is to prevent to accidentally kill processes that haven't been spawned by the compassomator (ie. old PID files after a system reboot). killing the wrong process is still possible, but a lot less likely. UPDATE: compassomator: stop watching on compassomator:clear; UPDATE: compassomator: make the cache warmer optional;

+++ Compassomator/ProcessManager.php
@@ -3,6 +3,7 @@
 namespace Asoc\CompassomatorBundle\Compassomator;
 
 use Symfony\Component\Finder\Finder;
+use Symfony\Component\Process\Process;
 
 class ProcessManager {
 
@@ -32,28 +33,27 @@ class ProcessManager {
 		return $this->runDir;
 	}
 
-	/**
-	 * @param $cmd
-	 * @param $name
-	 *
-	 * @return string
-	 */
-	public function run($cmd, $name) {
-		$cmd = sprintf('%s > %s & printf "%%u" $!', $cmd, $this->getLogFile($name));
-		$pid = shell_exec($cmd);
-		file_put_contents($this->getPidFile($name), $pid);
+	public function run(Process $process, $name) {
+		$cmd = $process->getCommandLine();
+		$pid = $this->runCommand($cmd, $name);
+
+		// the symfony process builder creates a commandline where every part is put into single quotes
+		// we extract the first part here, which is usually the actual command or program that is executed
+		// this should return something like: ruby or php
+		preg_match("'\w+'", $cmd, $matches);
+		$commandName = $matches[0];
+
+		// create a file where the lines are as following
+		// 1: PID
+		// 2: process name (ie. the actual command)
+		file_put_contents($this->getPidFile($name), sprintf("%d\n%s", $pid, $commandName));
 
 		return $pid;
 	}
 
-	/**
-	 * @param $name
-	 *
-	 * @return bool
-	 */
 	public function isRunning($name) {
 		$pidFile = $this->getPidFile($name);
-		return file_exists($pidFile);
+		return $this->isRunningByPidFile($pidFile);
 	}
 
 	/**
@@ -78,7 +78,10 @@ class ProcessManager {
 	 * @return int
 	 */
 	public function killByPidFile($pidFile) {
-		$pid = intval(file_get_contents($pidFile));
+		if(($pid = $this->isRunningByPidFile($pidFile)) === false) {
+			return false;
+		}
+
 		$cmd = sprintf('kill -9 %d > /dev/null 2>&1', $pid);
 		shell_exec($cmd);
 
@@ -98,7 +101,9 @@ class ProcessManager {
 
 		$pids = [];
 		foreach($remainingPids as $pidFile) {
-			$pids[] = $this->killByPidFile($pidFile);
+			if(($pid = $this->killByPidFile($pidFile)) !== false) {
+				$pids[] = $pid;
+			}
 		}
 
 		return $pids;
@@ -120,6 +125,43 @@ class ProcessManager {
 	 */
 	public function getPidFile($name) {
 		return sprintf('%s/%s.pid', $this->runDir, $name);
+	}
+
+	/**
+	 * @param string $cmd Full commandline to be executed
+	 * @param string $name Name (will be used for eg. the log file)
+	 *
+	 * @return string output of the command
+	 */
+	private function runCommand($cmd, $name) {
+		// wrap commandline so we can extract the PID and redirect the output to the log file
+		$cmd = sprintf('%s > %s & printf "%%u" $!', $cmd, $this->getLogFile($name));
+		return shell_exec($cmd);
+	}
+
+	/**
+	 * We return the PID here so we don't have to read the file multiple times just for the PID
+	 *
+	 * @param string $pidFile Path to the PID file
+	 *
+	 * @return bool|int false if the process is not running/doesn't match, the PID otherwise
+	 */
+	private function isRunningByPidFile($pidFile) {
+		if(!file_exists($pidFile)) {
+			return false;
+		}
+
+		$info = file($pidFile, FILE_IGNORE_NEW_LINES | FILE_SKIP_EMPTY_LINES);
+		$pid = $info[0];
+		$commandName = $info[1];
+
+		// we could also check for exit code
+		$result = trim(shell_exec(sprintf('ps -p %d -o comm=', $pid)));
+		if(strlen($result) !== 0 && $result === $commandName) {
+			return intval($pid);
+		}
+
+		return false;
 	}
 
 }
